Move implementation of create-drafts to new commands file
